Workaround samba "Invalid parameter" MIC bug (#226)
Disables sending of a SPNEGO MIC when guest authentication is requested.
(This cannot be verified anyways as there are no share credentials).

A range of samba versions has a bug that causes a "Invalid parameter"
error when a MIC is present and no key is available (invalid
credentials).

Also maps the invalid parameter error to a SmbAuthException to
permit proper handling and possibly fallback.