Try next DFS referral if fail to find link referral
While getReferralTransport() deals with the use case of a failed
connection to the referred server, it does not address a successful
connection that subsequently fails while calling getReferral.

There are probably a number of ways this could occur, but in our
case one set of servers where failing with
SmbUnsupportedOperationExceptions during getReferral, due to
misconfig. In the prior logic, this would cause an immediate
failure to find the link, rather than trying the other referrals
received from the getRootReferral.

The new logic follows the pattern used in getReferralTransport to
loop until we either succeed and break out or follow the referrals
all the way back to the first option.